requirements.txt
streamlit>=1.28.0,<2.0.0
google-generativeai>=0.3.0,<0.4.0
pypdf2>=3.0.0,<4.0.0
python-docx>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
wordcloud>=1.9.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
fpdf2>=2.7.0,<3.0.0
pytest>=7.0.0,<8.0.0